One char fix for the Wii menu and the cdb.vff issue as well as issue 1011, also fixed the Post build event, now copies files and directory if they're more recent without asking...
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3401 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoOGL/Src/Render.cpp
@@ -1305,7 +1305,7 @@ void Renderer::DrawDebugText()
 		OSDTime = timeGetTime() + 3000;
 		OSDChoice = -OSDChoice;
 	}
-	if (OSDTime > timeGetTime() && g_Config.bEFBCopyDisableHotKey)
+	if ((u32)OSDTime > timeGetTime() && g_Config.bEFBCopyDisableHotKey)
 	{
 		std::string T1 = "";
 		std::string T2 = "";
